+++ openfisca_us/data/datasets/cps/__init__.py
@@ -1,0 +1,2 @@
+from openfisca_us.data.datasets.cps.cps import CPS
+from openfisca_us.data.datasets.cps.raw_cps import RawCPS