+++ openfisca_us/variables/irs/credits/ctc/refundable.py
@@ -5,7 +5,7 @@ class ctc_refundable_individual_maximum(Variable):
     value_type = float
     entity = Person
     label = "Maximum refundable CTC"
-    unit = "currency-USD"
+    unit = USD
     documentation = "The maximum refundable CTC for this person."
     definition_period = YEAR
     reference = (
@@ -34,7 +34,7 @@ class ctc_limiting_tax_liability(Variable):
     value_type = float
     entity = TaxUnit
     label = "CTC-limiting tax liability"
-    unit = "currency-USD"
+    unit = USD
     documentation = "The tax liability used to determine the maximum amount of the non-refundable CTC."
     definition_period = YEAR
 
@@ -54,19 +54,33 @@ class refundable_ctc(Variable):
     value_type = float
     entity = TaxUnit
     label = "Refundable CTC"
-    unit = "currency-USD"
+    unit = USD
     documentation = "The portion of the Child Tax Credit that is refundable."
     definition_period = YEAR
     reference = "https://www.law.cornell.edu/uscode/text/26/24#d"
 
     def formula(tax_unit, period, parameters):
+
+        # This line corresponds to "the credit which would be allowed under this section [the CTC section]"
+        # without regard to this subsection [the refundability section] and the limitation under
+        # section 26(a) [the section that limits the amount of the non-refundable CTC to tax liability].
+        # This is the full CTC. This is then limited to the maximum refundable amount per child as per the
+        # TCJA provision.
+
+        ctc = parameters(period).irs.credits.ctc
+
+        total_ctc = tax_unit("ctc", period)
         maximum_refundable_ctc = min_(
             aggr(tax_unit, period, ["ctc_refundable_individual_maximum"]),
-            tax_unit("ctc", period),
+            total_ctc,
         )
 
-        liability = tax_unit("ctc_limiting_tax_liability", period)
-        ctc = parameters(period).irs.credits.ctc
+        # The other part of the "lesser of" statement is: "the amount by which [the non-refundable CTC]
+        # would increase if [tax liability] increased by tax_increase", where tax_increase is the greater of:
+        # - the phase-in amount
+        # - social security tax minus the EITC
+        # First, we find tax_increase:
+
         earnings = aggr(tax_unit, period, ["earned_income"])
         earnings_over_threshold = max_(
             0, earnings - ctc.refundable.phase_in.threshold
@@ -74,6 +88,7 @@ class refundable_ctc(Variable):
         relevant_earnings = (
             earnings_over_threshold * ctc.refundable.phase_in.rate
         )
+
         # Compute "social security taxes" as defined in the US Code for the ACTC.
         # This includes OASDI and Medicare payroll taxes, as well as half
         # of self-employment taxes.
@@ -95,10 +110,18 @@ class refundable_ctc(Variable):
         eitc = tax_unit("eitc", period)
         social_security_excess = max_(0, social_security_tax - eitc)
 
-        phased_in_amount = max_(relevant_earnings, social_security_excess)
+        tax_increase = max_(relevant_earnings, social_security_excess)
 
-        capped_phase_in = min_(phased_in_amount, liability)
-        return min_(capped_phase_in, maximum_refundable_ctc)
+        limiting_tax = tax_unit("ctc_limiting_tax_liability", period)
+        ctc_capped_by_tax = min_(total_ctc, limiting_tax)
+        ctc_capped_by_increased_tax = min_(
+            total_ctc, limiting_tax + tax_increase
+        )
+        amount_ctc_would_increase = (
+            ctc_capped_by_increased_tax - ctc_capped_by_tax
+        )
+
+        return min_(maximum_refundable_ctc, amount_ctc_would_increase)
 
     def formula_2021(tax_unit, period, parameters):
         return aggr(tax_unit, period, ["ctc_refundable_individual_maximum"])
@@ -110,7 +133,7 @@ class non_refundable_ctc(Variable):
     value_type = float
     entity = TaxUnit
     label = "Non-refundable CTC"
-    unit = "currency-USD"
+    unit = USD
     documentation = (
         "The portion of the Child Tax Credit that is not refundable."
     )